+++ tests/io/test_raster_io.py
@@ -4,6 +4,7 @@ For the full list of contributors, see the CREDITS file in the root directory of
 
 This source code is licensed under the MIT license, see the LICENSE file in the root directory of this source tree.
 """
+
 from __future__ import annotations
 
 import copy
@@ -15,13 +16,11 @@ import tempfile
 import warnings
 from typing import Any
 
-import boto3
 import numpy as np
 import pytest
 import rasterio
 from conftest import TEST_EOPATCH_PATH
 from fs.errors import ResourceNotFound
-from moto import mock_s3
 from numpy.testing import assert_array_equal
 
 from sentinelhub import CRS, BBox, read_data
@@ -32,17 +31,6 @@ from eolearn.core.exceptions import EORuntimeWarning, TemporalDimensionWarning
 from eolearn.io import ExportToTiffTask, ImportFromTiffTask
 
 logging.basicConfig(level=logging.DEBUG)
-
-BUCKET_NAME = "mocked-test-bucket"
-PATH_ON_BUCKET = f"s3://{BUCKET_NAME}/some-folder"
-
-
-@pytest.fixture(autouse=True)
-def _create_s3_bucket_fixture():
-    with mock_s3():
-        s3resource = boto3.resource("s3", region_name="eu-central-1")
-        s3resource.create_bucket(Bucket=BUCKET_NAME, CreateBucketConfiguration={"LocationConstraint": "eu-central-1"})
-        yield
 
 
 @pytest.fixture(autouse=True)
@@ -146,13 +134,13 @@ def test_export_import(test_case, test_eopatch):
         feature = test_case.feature_type, test_case.name
 
         export_task = ExportToTiffTask(
-            feature, folder=tmp_dir_name, band_indices=test_case.bands, date_indices=test_case.times
+            feature, path=tmp_dir_name, band_indices=test_case.bands, date_indices=test_case.times
         )
         _execute_with_warning_control(export_task, test_case.warning, test_eopatch, filename=tmp_file_name)
 
         export_task = ExportToTiffTask(
             feature,
-            folder=tmp_dir_name,
+            path=tmp_dir_name,
             band_indices=test_case.bands,
             date_indices=test_case.times,
             crs="EPSG:4326",
@@ -161,7 +149,7 @@ def test_export_import(test_case, test_eopatch):
         _execute_with_warning_control(export_task, test_case.warning, test_eopatch, filename=tmp_file_name_reproject)
 
         import_task = ImportFromTiffTask(
-            feature, folder=tmp_dir_name, timestamp_size=test_case.get_expected_timestamp_size()
+            feature, path=tmp_dir_name, timestamp_size=test_case.get_expected_timestamp_size()
         )
 
         expected_raster = test_case.get_expected()
@@ -209,7 +197,7 @@ def test_export2tiff_wrong_format(bands, times, test_eopatch):
     with tempfile.TemporaryDirectory() as tmp_dir_name:
         tmp_file_name = "temp_file.tiff"
         task = ExportToTiffTask(
-            (FeatureType.DATA, "data"), folder=tmp_dir_name, band_indices=bands, date_indices=times, image_dtype=float
+            (FeatureType.DATA, "data"), path=tmp_dir_name, band_indices=bands, date_indices=times, image_dtype=float
         )
         with pytest.raises(ValueError):
             task.execute(test_eopatch, filename=tmp_file_name)
@@ -222,7 +210,7 @@ def test_export2tiff_wrong_feature(mocker, test_eopatch):
         tmp_file_name = "temp_file.tiff"
         feature = FeatureType.MASK_TIMELESS, "feature-not-present"
 
-        export_task = ExportToTiffTask(feature, folder=tmp_dir_name, fail_on_missing=False)
+        export_task = ExportToTiffTask(feature, path=tmp_dir_name, fail_on_missing=False)
         export_task(test_eopatch, filename=tmp_file_name)
 
         assert logging.Logger.warning.call_count == 1
@@ -233,7 +221,7 @@ def test_export2tiff_wrong_feature(mocker, test_eopatch):
             == "Feature (<FeatureType.MASK_TIMELESS: 'mask_timeless'>, 'feature-not-present') was not found in EOPatch"
         )
 
-        failing_export_task = ExportToTiffTask(feature, folder=tmp_dir_name, fail_on_missing=True)
+        failing_export_task = ExportToTiffTask(feature, path=tmp_dir_name, fail_on_missing=True)
         with pytest.raises(ValueError):
             failing_export_task(test_eopatch, filename=tmp_file_name)
 
@@ -261,7 +249,7 @@ def test_export2tiff_separate_timestamps(test_eopatch):
         full_path = os.path.join(tmp_dir_name, tmp_file_name_reproject)
         export_task = ExportToTiffTask(
             feature,
-            folder=full_path,
+            path=full_path,
             band_indices=test_case.bands,
             date_indices=test_case.times,
             crs="EPSG:4326",
@@ -274,14 +262,11 @@ def test_export2tiff_separate_timestamps(test_eopatch):
             assert os.path.exists(expected_path), f"Path {expected_path} does not exist"
 
 
-# The following is not a proper test of use_vsi parameter. A proper test would require loading from an S3 path however
-# moto is not able to mock that because GDAL VSIS is not using boto3.
-@pytest.mark.parametrize("use_vsi", [True])  # , False])
-def test_import_tiff_subset(test_eopatch, example_data_path, use_vsi):
+def test_import_tiff_subset(test_eopatch, example_data_path):
     path = os.path.join(example_data_path, "import-tiff-test1.tiff")
     mask_feature = FeatureType.MASK_TIMELESS, "TEST_TIF"
 
-    task = ImportFromTiffTask(mask_feature, path, use_vsi=use_vsi)
+    task = ImportFromTiffTask(mask_feature, path)
     task(test_eopatch)
 
     tiff_img = read_data(path)
@@ -321,18 +306,18 @@ def test_import_tiff_intersecting(test_eopatch, example_data_path):
     assert unique_values == [no_data_value], f"No data values should all be equal to {no_data_value}"
 
 
-@pytest.mark.skip()  # rasterio 1.3.8 and moto have some issues, wait for it to be resolved
 def test_timeless_feature(test_eopatch):
     feature = FeatureType.DATA_TIMELESS, "DEM"
     filename = "relative-path/my-filename.tiff"
 
-    export_task = ExportToTiffTask(feature, folder=PATH_ON_BUCKET)
-    import_task = ImportFromTiffTask(feature, folder=PATH_ON_BUCKET)
+    with tempfile.TemporaryDirectory() as tmp_dir_name:
+        export_task = ExportToTiffTask(feature, path=tmp_dir_name)
+        import_task = ImportFromTiffTask(feature, path=tmp_dir_name)
 
-    export_task.execute(test_eopatch, filename=filename)
-    new_eopatch = import_task.execute(test_eopatch, filename=filename)
+        export_task.execute(test_eopatch, filename=filename)
+        new_eopatch = import_task.execute(test_eopatch, filename=filename)
 
-    assert_array_equal(new_eopatch[feature], test_eopatch[feature])
+        assert_array_equal(new_eopatch[feature], test_eopatch[feature])
 
 
 def test_time_dependent_feature(test_eopatch):
@@ -343,8 +328,8 @@ def test_time_dependent_feature(test_eopatch):
     ]
 
     with tempfile.TemporaryDirectory() as tmp_dir_name:
-        export_task = ExportToTiffTask(feature, folder=tmp_dir_name)
-        import_task = ImportFromTiffTask(feature, folder=tmp_dir_name, timestamp_size=68)
+        export_task = ExportToTiffTask(feature, path=tmp_dir_name)
+        import_task = ImportFromTiffTask(feature, path=tmp_dir_name, timestamp_size=68)
 
         export_task(test_eopatch, filename=filename_export)
         new_eopatch = import_task(filename=filename_import)
@@ -365,8 +350,8 @@ def test_time_dependent_feature_with_timestamps(test_eopatch):
     filename = "relative-path/%Y%m%dT%H%M%S.tiff"
 
     with tempfile.TemporaryDirectory() as tmp_dir_name:
-        export_task = ExportToTiffTask(feature, folder=tmp_dir_name)
-        import_task = ImportFromTiffTask(feature, folder=tmp_dir_name)
+        export_task = ExportToTiffTask(feature, path=tmp_dir_name)
+        import_task = ImportFromTiffTask(feature, path=tmp_dir_name)
 
         export_task.execute(test_eopatch, filename=filename)
         new_eopatch = import_task(test_eopatch, filename=filename)
@@ -390,7 +375,7 @@ def test_export_import_sequence(no_data_value, data_type):
         filename = "test_seq.tiff"
         file_path = os.path.join(tmp_dir_name, filename)
         export_task = ExportToTiffTask(
-            feature=feature, folder=tmp_dir_name, band_indices=[0], no_data_value=no_data_value
+            feature=feature, path=tmp_dir_name, band_indices=[0], no_data_value=no_data_value
         )
         export_task.execute(eopatch=eopatch, filename=filename)
 
@@ -405,7 +390,7 @@ def test_export_import_sequence(no_data_value, data_type):
 
             assert_array_equal(tif_array.mask, no_data_arr)
 
-        import_task = ImportFromTiffTask(feature=feature, folder=tmp_dir_name, no_data_value=no_data_value)
+        import_task = ImportFromTiffTask(feature=feature, path=tmp_dir_name, no_data_value=no_data_value)
         new_eopatch = import_task.execute(filename=filename)
 
         assert_array_equal(eopatch[feature], new_eopatch[feature])